# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/public/orders/review": {
      "post": {
        "operationId": "PublicOrdersController_review",
        "parameters": [
          {
            "name": "X-Tenant-Host",
            "in": "header",
            "description": "Replaces the Host header only when NODE_ENV is local or development",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewOrderDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderReviewResponse"
                }
              }
            }
          }
        },
        "tags": [
          "public-orders"
        ]
      }
    },
    "/api/v1/public/orders": {
      "post": {
        "operationId": "PublicOrdersController_create",
        "parameters": [
          {
            "name": "X-Tenant-Host",
            "in": "header",
            "description": "Replaces the Host header only when NODE_ENV is local or development",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrderDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreatedOrderResponse"
                }
              }
            }
          }
        },
        "tags": [
          "public-orders"
        ]
      }
    },
    "/api/v1/public/orders/{trackingToken}": {
      "get": {
        "operationId": "PublicOrdersController_get",
        "parameters": [
          {
            "name": "trackingToken",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicOrderResponse"
                }
              }
            }
          }
        },
        "tags": [
          "public-orders"
        ]
      }
    },
...
    "/api/v1/admin/orders/test": {
      "post": {
        "operationId": "AdminOrdersController_test",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminOrderResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin-orders"
        ]
      }
    },
    "/api/v1/admin/orders": {
      "get": {
        "operationId": "AdminOrdersController_list",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "string",
                "enum": [
                  "NEW",
                  "ACCEPTED",
                  "IN_PREPARATION",
                  "READY",
                  "OUT_FOR_DELIVERY",
                  "DELIVERED",
                  "REJECTED",
                  "CANCELLED"
                ]
              }
            }
          },
          {
            "name": "from",
            "required": false,
            "in": "query",
            "schema": {
              "format": "date-time",
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": false,
            "in": "query",
            "schema": {
              "format": "date-time",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminOrderPageResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin-orders"
        ]
      }
    },
    "/api/v1/admin/orders/{id}": {
      "get": {
        "operationId": "AdminOrdersController_get",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminOrderDetailResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin-orders"
        ]
      }
    },
    "/api/v1/admin/orders/{id}/accept": {
      "post": {
        "operationId": "AdminOrdersController_accept",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminOrderResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin-orders"
        ]
      }
    },
    "/api/v1/admin/orders/{id}/reject": {
      "post": {
        "operationId": "AdminOrdersController_reject",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OptionalOrderNoteDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminOrderResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin-orders"
        ]
      }
    },
    "/api/v1/admin/orders/{id}/start-preparation": {
      "post": {
        "operationId": "AdminOrdersController_startPreparation",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminOrderResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin-orders"
        ]
      }
    },
...
    "/api/v1/admin/orders/{id}/cancel": {
      "post": {
        "operationId": "AdminOrdersController_cancel",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CancelOrderDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminOrderResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin-orders"
        ]
      }
    },
    "/api/v1/admin/orders/{id}/assign-courier": {
      "post": {
        "operationId": "AdminOrdersController_assignCourier",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignCourierDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminOrderResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin-orders"
        ]
      }
    },
    "/api/v1/admin/orders/{id}/dispatch": {
      "post": {
        "operationId": "AdminOrdersController_dispatch",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminOrderResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin-orders"
        ]
      }
    },
...
    "/api/v1/admin/orders/{id}/complete-pickup": {
      "post": {
        "operationId": "AdminOrdersController_completePickup",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminOrderResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin-orders"
        ]
      }
    },
...
    "/api/v1/courier/orders/{id}": {
      "get": {
        "operationId": "CourierOrdersController_get",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CourierOrderDetailResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "courier-orders"
        ]
      }
    },
...
  },
  "info": {
    "title": "ciadelivery",
    "description": "ciadelivery HTTP API",
    "version": "0.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "OrderAddressDto": {
        "type": "object",
        "properties": {
          "line": {
            "type": "string"
          },
          "number": {
            "type": "string"
          },
          "district": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "state": {
            "type": "string",
            "example": "SP"
          },
          "postalCode": {
            "type": "string",
            "example": "01001000"
          },
          "complement": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "line",
          "number",
          "district",
          "city",
          "state",
          "postalCode"
        ]
      },
...
      "ReviewOrderDto": {
        "type": "object",
        "properties": {
          "fulfillment": {
            "type": "string",
            "enum": [
              "DELIVERY",
              "PICKUP"
            ]
          },
          "address": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/OrderAddressDto"
              }
            ]
          },
          "paymentMethodCode": {
            "type": "string",
            "enum": [
              "CASH",
              "CARD_ON_DELIVERY",
              "PIX_MANUAL",
              "PAY_ON_PICKUP",
              "OTHER"
            ]
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreateOrderItemDto"
            }
          }
        },
        "required": [
          "fulfillment",
          "paymentMethodCode",
          "items"
        ]
      },
...
      "OrderReviewResponse": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderReviewItemResponse"
            }
          },
          "subtotalCents": {
            "type": "number"
          },
          "deliveryFeeCents": {
            "type": "number"
          },
          "totalCents": {
            "type": "number"
          },
          "paymentLabel": {
            "type": "string"
          },
          "paymentInstructions": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "items",
          "subtotalCents",
          "deliveryFeeCents",
          "totalCents",
          "paymentLabel",
          "paymentInstructions"
        ]
      },
...
      "CreateOrderDto": {
        "type": "object",
        "properties": {
          "customer": {
            "$ref": "#/components/schemas/OrderCustomerDto"
          },
          "fulfillment": {
            "type": "string",
            "enum": [
              "DELIVERY",
              "PICKUP"
            ]
          },
          "address": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/OrderAddressDto"
              }
            ]
          },
          "paymentMethodCode": {
            "type": "string",
            "enum": [
              "CASH",
              "CARD_ON_DELIVERY",
              "PIX_MANUAL",
              "PAY_ON_PICKUP",
              "OTHER"
            ]
          },
          "notes": {
            "type": "string",
            "nullable": true,
            "maxLength": 280
          },
          "consents": {
            "$ref": "#/components/schemas/OrderConsentsDto"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreateOrderItemDto"
            }
          }
        },
        "required": [
          "customer",
          "fulfillment",
          "paymentMethodCode",
          "consents",
          "items"
        ]
      },
      "CreatedOrderResponse": {
        "type": "object",
        "properties": {
          "orderId": {
            "type": "string"
          },
          "orderNumber": {
            "type": "number"
          },
          "status": {
            "type": "string",
            "enum": [
              "NEW"
            ]
          },
          "totalCents": {
            "type": "number"
          },
          "trackingToken": {
            "type": "string"
          },
          "trackingPath": {
            "type": "string"
          }
        },
        "required": [
          "orderId",
          "orderNumber",
          "status",
          "totalCents",
          "trackingToken",
          "trackingPath"
        ]
      },
...
      "PublicOrderResponse": {
        "type": "object",
        "properties": {
          "orderId": {
            "type": "string"
          },
          "orderNumber": {
            "type": "number"
          },
          "status": {
            "type": "string",
            "enum": [
              "NEW",
              "ACCEPTED",
              "IN_PREPARATION",
              "READY",
              "OUT_FOR_DELIVERY",
              "DELIVERED",
              "REJECTED",
              "CANCELLED"
            ]
          },
          "fulfillment": {
            "type": "string",
            "enum": [
              "DELIVERY",
              "PICKUP"
            ]
          },
          "paymentMethodCode": {
            "type": "string"
          },
          "paymentLabel": {
            "type": "string"
          },
          "paymentInstructions": {
            "type": "string",
            "nullable": true
          },
          "customerName": {
            "type": "string"
          },
          "customerPhone": {
            "type": "string"
          },
          "address": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/PublicOrderAddressResponse"
              }
            ]
          },
          "notes": {
            "type": "string",
            "nullable": true
          },
          "subtotalCents": {
            "type": "number"
          },
          "deliveryFeeCents": {
            "type": "number"
          },
          "totalCents": {
            "type": "number"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PublicOrderItemResponse"
            }
          },
          "history": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PublicOrderHistoryResponse"
            }
          }
        },
        "required": [
          "orderId",
          "orderNumber",
          "status",
          "fulfillment",
          "paymentMethodCode",
          "paymentLabel",
          "paymentInstructions",
          "customerName",
          "customerPhone",
          "address",
          "notes",
          "subtotalCents",
          "deliveryFeeCents",
          "totalCents",
          "createdAt",
          "items",
          "history"
        ]
      },
...
      "AdminOrderResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "orderNumber": {
            "type": "number"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "totalCents": {
            "type": "number"
          },
          "status": {
            "type": "string",
            "enum": [
              "NEW",
              "ACCEPTED",
              "IN_PREPARATION",
              "READY",
              "OUT_FOR_DELIVERY",
              "DELIVERED",
              "REJECTED",
              "CANCELLED"
            ]
          },
          "fulfillment": {
            "type": "string",
            "enum": [
              "DELIVERY",
              "PICKUP"
            ]
          },
          "customerName": {
            "type": "string"
          },
          "source": {
            "type": "string",
            "enum": [
              "STOREFRONT",
              "WHATSAPP",
              "TEST"
            ]
          },
          "notes": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "id",
          "orderNumber",
          "createdAt",
          "totalCents",
          "status",
          "fulfillment",
          "customerName",
          "source",
          "notes"
        ]
      },
...
      "AdminOrderPageResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AdminOrderResponse"
            }
          },
          "meta": {
            "$ref": "#/components/schemas/AdminOrderPageMeta"
          }
        },
        "required": [
          "data",
          "meta"
        ]
      },
      "AdminOrderDetailResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "orderNumber": {
            "type": "number"
          },
          "status": {
            "type": "string",
            "enum": [
              "NEW",
              "ACCEPTED",
              "IN_PREPARATION",
              "READY",
              "OUT_FOR_DELIVERY",
              "DELIVERED",
              "REJECTED",
              "CANCELLED"
            ]
          },
          "fulfillment": {
            "type": "string",
            "enum": [
              "DELIVERY",
              "PICKUP"
            ]
          },
          "source": {
            "type": "string",
            "enum": [
              "STOREFRONT",
              "WHATSAPP",
              "TEST"
            ]
          },
          "paymentMethodCode": {
            "type": "string"
          },
          "paymentLabel": {
            "type": "string"
          },
          "paymentInstructions": {
            "type": "string",
            "nullable": true
          },
          "customerName": {
            "type": "string"
          },
          "customerPhone": {
            "type": "string"
          },
          "address": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/PublicOrderAddressResponse"
              }
            ]
          },
          "notes": {
            "type": "string",
            "nullable": true
          },
          "subtotalCents": {
            "type": "number"
          },
          "deliveryFeeCents": {
            "type": "number"
          },
          "totalCents": {
            "type": "number"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PublicOrderItemResponse"
            }
          },
          "history": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PublicOrderHistoryResponse"
            }
          }
        },
        "required": [
          "id",
          "orderNumber",
          "status",
          "fulfillment",
          "source",
          "paymentMethodCode",
          "paymentLabel",
          "paymentInstructions",
          "customerName",
          "customerPhone",
          "address",
          "notes",
          "subtotalCents",
          "deliveryFeeCents",
          "totalCents",
          "createdAt",
          "items",
          "history"
        ]
      },
      "OptionalOrderNoteDto": {
        "type": "object",
        "properties": {
          "note": {
            "type": "object",
            "nullable": true,
            "maxLength": 280
          }
        }
      },
      "CancelOrderDto": {
        "type": "object",
        "properties": {
          "note": {
            "type": "string",
            "maxLength": 280
          }
        },
        "required": [
          "note"
        ]
      },
      "AssignCourierDto": {
        "type": "object",
        "properties": {
          "courierId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "courierId"
        ]
      },
...
      "CourierOrderDetailResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "orderNumber": {
            "type": "number"
          },
          "status": {
            "type": "string",
            "enum": [
              "NEW",
              "ACCEPTED",
              "IN_PREPARATION",
              "READY",
              "OUT_FOR_DELIVERY",
              "DELIVERED",
              "REJECTED",
              "CANCELLED"
            ]
          },
          "totalCents": {
            "type": "number"
          },
          "address": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/CourierAddressResponse"
              }
            ]
          },
          "customerPhone": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "orderNumber",
          "status",
          "totalCents",
          "address",
          "customerPhone"
        ]
      },
...
    }
  }
}